% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{autoplot.posterior}
\alias{autoplot.posterior}
\alias{autoplot.perf_mod}
\title{Visualize the Posterior Distributions of Model Statistics}
\usage{
\method{autoplot}{posterior}(object, ...)

\method{autoplot}{perf_mod}(object, ...)
}
\arguments{
\item{object}{An object produced by \code{\link[=perf_mod]{perf_mod()}} or  \code{\link[=tidy.perf_mod]{tidy.perf_mod()}}.}

\item{...}{Options passed to \code{geom_line(stat = "density", ...)}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object using \code{geom_line(stat = "density", ...)}
for the posteriors.
}
\description{
A simple plot of posterior distributions.
}
\examples{
data(ex_objects)
autoplot(posterior_samples)
}
