% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_resample.R
\name{Bayes_resample}
\alias{Bayes_resample}
\alias{Bayes_resample.rset}
\title{Bayesian Analysis of Resampling Statistics}
\usage{
Bayes_resample(object, ...)

\method{Bayes_resample}{rset}(object, transform = no_trans,
  hetero_var = FALSE, ...)
}
\arguments{
\item{object}{A data frame or and \code{rset} object (such as
\code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}) containing the \code{id} column(s) and at least
two numeric columns of model performance statistics (e.g.
accuracy).}

\item{...}{Additonal arguments to pass to \code{\link[rstanarm:stan_glmer]{rstanarm::stan_glmer()}}
such as \code{verbose}, \code{prior}, \code{seed}, etc.}

\item{transform}{An named list of transformation and inverse
transformation fuctions. See \code{\link[=logit_trans]{logit_trans()}} as an example.}

\item{hetero_var}{A logical; if \code{TRUE}, then different
variances are estimated for each model group. Otherwise, the
same variance is used for each group. Estimating heterogeneous
variances may slow or prevent convergence.}
}
\value{
An object of class \code{Bayes_resample}.
}
\description{
These functions can be used to process and analyze matched
resampling statistics from different models using a Bayesian
generalized linear model with effects for the model and the
resamples.
}
