% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{ggplot.posterior}
\alias{ggplot.posterior}
\title{Visualize the Posterior Distirbutions of Model Statistics}
\usage{
\method{ggplot}{posterior}(data, mapping = NULL, ..., environment = NULL,
  reorder = TRUE)
}
\arguments{
\item{data}{An object produced by \code{\link[=tidy.Bayes_resample]{tidy.Bayes_resample()}}.}

\item{mapping, ..., environment}{Not currently used.}

\item{reorder}{A logical; should the \code{model} column be reordered
by the average of the posterior distirbution?}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
A simple violin plot is created by the function.
}
