% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{ggplot.posterior_diff}
\alias{ggplot.posterior_diff}
\title{Visualize the Posterior Distributions of Model Differences}
\usage{
\method{ggplot}{posterior_diff}(data, mapping = NULL, ...,
  environment = NULL, size = 0)
}
\arguments{
\item{data}{An object produced by \code{\link[=contrast_models]{contrast_models()}}.}

\item{mapping, ..., environment}{Not currently used.}

\item{size}{The size of an effective difference. For example, a
5\% increase in accuracy between two models might be considered a
"real" difference.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object using \code{geom_density}
facetted by the models being contrasted (when there are 2 or
more contrasts).
}
\description{
A density is created for each contrast in a facetted grid.
}
\examples{
# Example from "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

file <- system.file("examples", "glm_v_nnet.RData", package = "tidyposterior")
load(file)

ggplot(glm_v_nnet, size = 0.02) + theme_bw()
 
}
