% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyposterior-package.R
\docType{package}
\name{tidyposterior-package}
\alias{tidyposterior}
\alias{tidyposterior-package}
\title{tidyposterior: Bayesian Analysis to Compare Models using Resampling Statistics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayesian analysis used here to answer the question: "when looking at resampling results, are the differences between models 'real'?" To answer this, a model can be created were the performance statistic is the resampling statistics (e.g. accuracy or RMSE). These values are explained by the model types. In doing this, we can get parameter estimates for each model's affect on performance and make statistical (and practical) comparisons between models. The methods included here are similar to Benavoli et al (2017) \url{https://jmlr.org/papers/v18/16-305.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidyposterior.tidymodels.org}
  \item \url{https://github.com/tidymodels/tidyposterior}
  \item Report bugs at \url{https://github.com/tidymodels/tidyposterior/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
