% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_diff-compat.R
\name{vec_restore.posterior_diff}
\alias{vec_restore.posterior_diff}
\alias{vec_proxy.posterior_diff}
\alias{vec_ptype2.posterior_diff.posterior_diff}
\alias{vec_ptype2.posterior_diff.tbl_df}
\alias{vec_ptype2.tbl_df.posterior_diff}
\alias{vec_ptype2.posterior_diff.data.frame}
\alias{vec_ptype2.data.frame.posterior_diff}
\alias{vec_cast.posterior_diff.posterior_diff}
\alias{vec_cast.posterior_diff.tbl_df}
\alias{vec_cast.tbl_df.posterior_diff}
\alias{vec_cast.posterior_diff.data.frame}
\alias{vec_cast.data.frame.posterior_diff}
\title{Extra methods for the \code{posterior_diff} class to work with dplyr verbs}
\usage{
vec_restore.posterior_diff(x, to, ...)

vec_proxy.posterior_diff(x, ...)

vec_ptype2.posterior_diff.posterior_diff(x, y, ..., x_arg = "", y_arg = "")

vec_ptype2.posterior_diff.tbl_df(x, y, ..., x_arg = "", y_arg = "")

vec_ptype2.tbl_df.posterior_diff(x, y, ..., x_arg = "", y_arg = "")

vec_ptype2.posterior_diff.data.frame(x, y, ..., x_arg = "", y_arg = "")

vec_ptype2.data.frame.posterior_diff(x, y, ..., x_arg = "", y_arg = "")

vec_cast.posterior_diff.posterior_diff(x, to, ..., x_arg = "", to_arg = "")

vec_cast.posterior_diff.tbl_df(x, to, ..., x_arg = "", to_arg = "")

vec_cast.tbl_df.posterior_diff(x, to, ..., x_arg = "", to_arg = "")

vec_cast.posterior_diff.data.frame(x, to, ..., x_arg = "", to_arg = "")

vec_cast.data.frame.posterior_diff(x, to, ..., x_arg = "", to_arg = "")
}
\description{
Objects with class \code{posterior_diff} are defined to be tibbles with required
columns \code{difference} (numeric) and character columns \code{model_1}, \code{model_2},
and \code{contrast}. If operations on these objects break those rules, they are
down-cast to basic tibbles.
}
\keyword{internal}
