% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{ggplot.posterior}
\alias{ggplot.posterior}
\title{Visualize the Posterior Distributions of Model Statistics}
\usage{
\method{ggplot}{posterior}(data, mapping = NULL, ..., environment = NULL, reorder = TRUE)
}
\arguments{
\item{data}{An object produced by \code{\link[=tidy.perf_mod]{tidy.perf_mod()}}.}

\item{mapping, ..., environment}{Not currently used.}

\item{reorder}{A logical; should the \code{model} column be reordered
by the average of the posterior distribution?}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object using
\code{\link[ggplot2:geom_violin]{ggplot2::geom_violin()}} for the posteriors.
}
\description{
A simple violin plot is created by the function.
}
\details{
\lifecycle{deprecated}
}
\examples{
data(ex_objects)
library(ggplot2)
ggplot(posterior_samples)
}
