% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{tidy.Bayes_resample}
\alias{tidy.Bayes_resample}
\title{Extract Posterior Distributions for Models}
\usage{
\method{tidy}{Bayes_resample}(x, ...)
}
\arguments{
\item{x}{An object from \code{\link[=Bayes_resample]{Bayes_resample()}}}

\item{...}{Not currently used}
}
\value{
A data frame with the additional class \code{"posterior"}
}
\description{
\code{tidy} can be used on an object produced by \code{\link[=Bayes_resample]{Bayes_resample()}}
to create a data frame with a column for the model name and
the posterior predictive distribution values.
}
